# backend/src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Avanzada API",
    "description": "REST API for managing support requests (students, staff, admin).",
    "contact": {
      "name": "Avanzada"
    },
    "license": {
      "name": "MIT"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/ai/suggest": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Suggest type and priority using AI",
        "description": "Suggests request type and priority from a free-text description. Suggestions are never auto-applied.",
        "operationId": "suggest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Suggestion generated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/channels": {
      "get": {
        "tags": [
          "Channels"
        ],
        "summary": "List channels",
        "description": "Returns all channels (e.g. WEB, EMAIL) that can be used when registering requests.",
        "operationId": "listChannels",
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "200": {
            "description": "Channels found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SuggestRequestDto": {
        "required": [
          "description"
        ],
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          }
        }
      },
      "SuggestResponseDto": {
        "type": "object",
        "properties": {
          "suggestedRequestTypeCode": {
            "type": "string"
          },
          "suggestedPriority": {
            "type": "string"
          },
          "available": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ChannelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Bearer token obtained via /api/auth/login",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}